requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
slowapi
pydantic
python-multipart